+++ Ray.Core/EventSourcing/ESGrain.cs
@@ -13,7 +13,7 @@ namespace Ray.Core.EventSourcing
         where S : class, IState<K>, new()
         where W : MessageWrapper
     {
-        UInt32 storageVersion;
+        Int64 storageVersion;
         protected S State
         {
             get;
